Cluster; K-Means; Dendrogram
The cluster.py depicts how to perform hierarchical clusters (Dendrogram) and k-means cluster. For dendrogram, the number of clusters is determined manually based on the threshold selected; For k-means cluster, the number of clusters is set ahead and the process stops when the old centroids and new centroids are converged.